requirements.txt
Django==4.2.20
djangorestframework==3.15.2
django-filter==25.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2

# Computation
geopy

# Infrastracture
uvicorn[standard]
psycopg2-binary>=2.9.1

# Documentation
drf-spectacular>=0.24.2
django-debug-toolbar
django-silk
